+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-05-18 18:18:40
-// Commit: feat(i18n): add bi-directional event bus handler for i18n
+// Updated at: 2026-05-18 20:24:40
+// Commit: docs(i18n): update readme and API contracts for i18n
 
 class i18nComponentState {
-  final String commitRef = '1267';
-  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
-  final DateTime timestamp = DateTime.parse('2026-05-18T18:18:40+05:00');
+  final String commitRef = '1271';
+  final String lastMessage = 'docs(i18n): update readme and API contracts for i18n';
+  final DateTime timestamp = DateTime.parse('2026-05-18T20:24:40+05:00');
 }